Python Script for Converting GPS Coordinates to a KML File
This script allows you to take a set of GPS coordinates and automatically generate a KML file, which can be viewed in geographic software like Google Earth. The script reads latitude and longitude pairs, creates corresponding placemarks, and saves the result as a .kml file. This is useful for visualizing and sharing geographical data on maps.